source/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2023.1
cryptography==42.0.3
defusedxml==0.7.1
distutils-pytest==0.2.1
Django==5.0.3
django-allauth==0.61.1
django-crispy-forms==2.1
django-filter==24.2
django-jalali-date==1.1.0
django-rosetta==0.10.0
exceptiongroup==1.2.1
idna==3.6
iniconfig==2.0.0
install==1.3.5
jalali_core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==1.5.0
polib==1.2.0
pycparser==2.21
PyJWT==2.8.0
pytest==8.2.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==70.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
